requirements.txt
tree-sitter
tree-sitter-python
tree-sitter-javascript
langchain
langchain-community
langchain-huggingface
langgraph
langchain-text-splitters
PyGithub
pytest
pytest-cov
pytest-mock
streamlit
qdrant-client
sentence-transformers
transformers
torch
accelerate
python-dotenv
fastapi
uvicorn
httpx
